requirements.txt
pytest
pywinauto
openpyxl
pandas
requests
Pillow
